fix(#289): move birth kwargs test to test_soul.py, address review
- Moved test_birth_unknown_kwargs_logs_warning from test_contradiction_pipeline.py
  to test_soul.py where it belongs with soul lifecycle tests
- Item 5 (84 brittle Rich assertions) deferred to #289 follow-up
